+++ server.js
@@ -56,6 +56,18 @@ app.prepare().then(() => {
     return app.render(req, res, '/', req.query)
   })
 
+  server.get('/gLogin', (req, res) => {
+    req.query.modal = 'login'
+    req.query.socialProvider = 'google'
+    return app.render(req, res, '/', req.query)
+  })
+
+  server.get('/fLogin', (req, res) => {
+    req.query.modal = 'login'
+    req.query.socialProvider = 'facebook'
+    return app.render(req, res, '/', req.query)
+  })
+
   server.get('/register', (req, res) => {
     req.query.modal = 'register'
     return app.render(req, res, '/', req.query)